Stop UI jumping when advisor switches messages
I couldn't figure out a better way to do this, it leaves some empty
space between the advisor section and the stats header, but that doesn't
seem *too* bad to me? I also increased the padding on the sidebar to
avoid the text overlapped the dark spots on the background image.